+++ Revit.Busy/Revit/App.cs
@@ -9,14 +9,17 @@ namespace Revit.Busy.Revit
     public class App : IExternalApplication
     {
         private static RibbonPanel ribbonPanel;
+        private static RibbonItem ribbonItem;
         public Result OnStartup(UIControlledApplication application)
         {
             ribbonPanel = application.CreatePanel("Revit.Busy");
-            ribbonPanel.CreatePushButton<Commands.Command>()
+            ribbonItem = ribbonPanel.CreatePushButton<Commands.Command>()
                 .SetLargeImage("/UIFrameworkRes;component/ribbon/images/revit.ico");
 
             RevitBusyControl.Initialize(application);
             RevitBusyControl.Control.PropertyChanged += RevitBusyControlPropertyChanged;
+
+            UpdateLargeImageBusy(ribbonItem, RevitBusyControl.Control);
 
             return Result.Succeeded;
         }
@@ -24,6 +27,9 @@ namespace Revit.Busy.Revit
         private void RevitBusyControlPropertyChanged(object sender, System.ComponentModel.PropertyChangedEventArgs e)
         {
             Console.WriteLine($"RevitBusyControl PropertyChanged {e.PropertyName} {RevitBusyControl.Control.IsRevitBusy}");
+
+            var control = sender as RevitBusyService;
+            UpdateLargeImageBusy(ribbonItem, control);
         }
 
         public Result OnShutdown(UIControlledApplication application)
@@ -32,6 +38,16 @@ namespace Revit.Busy.Revit
             RevitBusyControl.Control.PropertyChanged -= RevitBusyControlPropertyChanged;
             RevitBusyControl.Dispose();
             return Result.Succeeded;
+        }
+
+        private static void UpdateLargeImageBusy(RibbonItem ribbonItem, RevitBusyService control)
+        {
+            const string LargeImageIsBusy = "/UIFrameworkRes;component/ribbon/images/close.ico";
+            const string LargeImageNoBusy = "/UIFrameworkRes;component/ribbon/images/add.ico";
+            if (control.IsRevitBusy)
+                ribbonItem.SetLargeImage(LargeImageIsBusy);
+            else
+                ribbonItem.SetLargeImage(LargeImageNoBusy);
         }
     }