+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.1.2] / 2023-02-22
+### Features
+- `RevitBusyService` - start busy
+### Changed
+- `RevitBusyService` default `isRevitBusy = true`
+
 ## [0.1.1] / 2022-12-12
 ### Changed
 - Change timer busy to 1 second
@@ -17,5 +23,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [vNext]: ../../compare/1.0.0...HEAD
 [1.0.0]: ../../compare/1.0.0
+[0.1.2]: ../../compare/0.1.1...0.1.2
 [0.1.1]: ../../compare/0.1.0...0.1.1
 [0.1.0]: ../../compare/0.1.0